NO-ISSUE: Fix ObjectLikeKeys type error, add ObjectLikeFields (#4328)
* Refactor type, add ObjctLikeFields

* add Flatten to tsutils pkg
