+++ packages/server-boilerplate/package.json
@@ -28,6 +28,7 @@
     "@tupaia/api-client": "3.1.0",
     "@tupaia/auth": "1.0.0",
     "@tupaia/database": "1.0.0",
+    "@tupaia/tsutils": "1.0.0",
     "@tupaia/utils": "1.0.0",
     "api-error-handler": "^1.0.0",
     "body-parser": "^1.18.3",
