+++ packages/tsutils/src/types.ts
@@ -24,3 +24,33 @@ export type NullableKeys<T> = { [K in keyof T]-?: null | T[K] extends T[K] ? K :
  */
 export type NullableKeysToOptional<T> = Pick<T, NonNullKeys<T>> &
   { [P in NullableKeys<T>]?: Exclude<T[P], null> };
+
+/**
+ * Extracts keys that have object-like values from type T
+ */
+export type ObjectLikeKeys<T> = {
+  [K in keyof T]: T[K] extends Record<string, unknown> ? K : never;
+}[keyof T];
+
+// Extracts fields that have object-like values from type T
+export type ObjectLikeFields<T> = {
+  [K in keyof T]: T[K] extends Record<string, unknown> ? T[K] : never;
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type UnionToIntersection<U> = (U extends any ? (k: U) => void : never) extends (k: infer I) => void
+  ? I
+  : never;
+
+// Flattens nested object to shallow object with keys joined by J
+// eg. Flatten<{ cat: { cute: true } }, '_is_'> => { cat_is_cute: true }
+export type Flatten<
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  T extends Record<string, Record<string, any>>,
+  J extends string = '.',
+  K extends keyof T & string = keyof T & string
+> = UnionToIntersection<
+  {
+    [V in K]: { [field in keyof T[V] & string as `${V}${J}${field}`]: T[V][field] };
+  }[K]
+>;
